requirements.txt
ansible==1.9.3
appdirs~=1.4
arrow~=0.5
click~=3.3
colorama~=0.3
flake8~=2.3
github3.py~=0.9
GitPython~=0.3
Jinja2~=2.7
netaddr~=0.7
nose~=1.3
ordereddict==1.1
prompt-toolkit==0.31
python-slugify~=1.0
raven~=5.2.0
semver~=2.0
sh~=1.11
Sphinx~=1.2
watchdog~=0.8
git+https://github.com/christopherobin/python-subprocess32.git#egg=subprocess32
git+https://github.com/dominis/ansible-shell.git#egg=ansible-shell
git+https://github.com/willthames/ansible-lint.git#egg=ansible-lint
